Crate re_query

Provide query-centric access to the [re_arrow_store].

Feature flags

polars — Polars support

Modules

dataframe_util

entity_view 🔒

query 🔒

range 🔒

util 🔒

visit

Visit the primary and joined components of an EntityView

Structs

ComponentWithInstances

A type-erased array of [Component] values and the corresponding [InstanceKey] keys.

EntityView

A view of an entity at a particular point in time returned by crate::get_component_with_instances

Enums

QueryError

Functions

get_component_with_instances

Retrieves a ComponentWithInstances from the [DataStore].

query_entity_with_primary

Retrieve an EntityView from the DataStore

query_primary_with_history

Either dispatch to query_entity_with_primary or range_entity_with_primary depending on whether ExtraQueryHistory is set.

range_entity_with_primary

Iterates over the rows of any number of components and their respective cluster keys, all from the single point-of-view of the primary component, returning an iterator of EntityViews.

Type Definitions

Result