re_query::clamped_zip
Function clamped_zip_1x3

pub fn clamped_zip_1x3<R0, O0, O1, O2, D0, D1, D2>(
    r0: R0,
    o0: O0,
    o0_default_fn: D0,
    o1: O1,
    o1_default_fn: D1,
    o2: O2,
    o2_default_fn: D2,
) -> ClampedZip1x3<R0::IntoIter, O0::IntoIter, O1::IntoIter, O2::IntoIter, D0, D1, D2> 
where
    R0: IntoIterator,
    O0: IntoIterator,
    O0::Item: Clone,
    O1: IntoIterator,
    O1::Item: Clone,
    O2: IntoIterator,
    O2::Item: Clone,
    D0: Fn() -> O0::Item,
    D1: Fn() -> O1::Item,
    D2: Fn() -> O2::Item,

Returns a new ClampedZip1x3 iterator.

The number of elements in a clamped zip iterator corresponds to the number of elements in the shortest of its required iterators (r0).

Optional iterators (o0, o1, o2) will repeat their latest values if they happen to be too short to be zipped with the shortest of the required iterators.

If an optional iterator is not only too short but actually empty, its associated default function (o0_default_fn, o1_default_fn, o2_default_fn) will be executed and the resulting value repeated as necessary.