Function re_query::clamped_zip::generated::clamped_zip_2x2

pub fn clamped_zip_2x2<R0, R1, O0, O1, D0, D1>(
    r0: R0,
    r1: R1,
    o0: O0,
    o0_default_fn: D0,
    o1: O1,
    o1_default_fn: D1
) -> ClampedZip2x2<R0::IntoIter, R1::IntoIter, O0::IntoIter, O1::IntoIter, D0, D1> 
where
    R0: IntoIterator,
    R1: IntoIterator,
    O0: IntoIterator,
    O0::Item: Clone,
    O1: IntoIterator,
    O1::Item: Clone,
    D0: Fn() -> O0::Item,
    D1: Fn() -> O1::Item,

Returns a new ClampedZip2x2 iterator.

The number of elements in a clamped zip iterator corresponds to the number of elements in the shortest of its required iterators (r0, r1).

Optional iterators (o0, o1) will repeat their latest values if they happen to be too short to be zipped with the shortest of the required iterators.

If an optional iterator is not only too short but actually empty, its associated default function (o0_default_fn, o1_default_fn) will be executed and the resulting value repeated as necessary.