Struct re_query::entity_view::EntityView

pub struct EntityView<Primary: Component> {
    pub(crate) row_id: RowId,
    pub(crate) primary: ComponentWithInstances,
    pub(crate) components: BTreeMap<ComponentName, ComponentWithInstances>,
    pub(crate) phantom: PhantomData<Primary>,
}

A view of an entity at a particular point in time returned by crate::get_component_with_instances

EntityView has a special primary Component which determines the length of an entity batch. When iterating over individual components, they will be implicitly joined onto the primary component using instance keys.

Fields
row_id: RowIdprimary: ComponentWithInstancescomponents: BTreeMap<ComponentName, ComponentWithInstances>phantom: PhantomData<Primary>
Implementations

impl<Primary> EntityView<Primary>where Primary: Component,

pub fn num_instances(&self) -> usize

pub fn row_id(&self) -> RowId

impl<'a, Primary> EntityView<Primary>where Primary: Component + 'a, &'a Primary: Into<Cow<'a, Primary>>,

pub fn iter_instance_keys(&self) -> impl Iterator<Item = InstanceKey> + '_

Iterate over the instance keys

pub fn iter_primary(&self) -> Result<impl Iterator<Item = Option<Primary>> + '_>

Iterate over the primary component values.

pub fn iter_primary_flattened(&self) -> impl Iterator<Item = Primary> + '_

Iterate over the flattened list of primary component values if any.

pub fn has_component<C: Component>(&self) -> bool

Check if the entity has a component and its not empty

pub fn iter_component<'b, C: Component + 'b>( &'b self ) -> Result<impl Iterator<Item = Option<C>> + '_>

Iterate over the values of a Component.

Always produces an iterator of length self.primary.len()

pub fn from_native(c0: (&'a [InstanceKey], &'a [Primary])) -> Self

Helper function to produce an EntityView from rust-native field_types

pub fn from_native2<C>( primary: (&'a [InstanceKey], &'a [Primary]), component: (&'a [InstanceKey], &'a [C]) ) -> Selfwhere C: Component + 'a, &'a C: Into<Cow<'a, C>>,

Helper function to produce an EntityView from rust-native field_types

impl<'a, Primary> EntityView<Primary>where Primary: Component, &'a Primary: Into<Cow<'a, Primary>> + 'a,

pub fn visit1(&self, visit: impl FnMut(InstanceKey, Primary)) -> Result<()>

Visit the primary component of an EntityView. See crate::visit

pub fn visit2<C1>( &self, visit: impl FnMut(InstanceKey, Primary, Option<C1>) ) -> Result<()>where C1: Component + Component, &'a C1: Into<Cow<'a, C1>> + 'a,

Visit the primary component of an EntityView. See crate::visit

pub fn visit3<C1, C2>( &self, visit: impl FnMut(InstanceKey, Primary, Option<C1>, Option<C2>) ) -> Result<()>where C1: Component + Component, C2: Component + Component, &'a C1: Into<Cow<'a, C1>> + 'a, &'a C2: Into<Cow<'a, C2>> + 'a,

Visit the primary component of an EntityView. See crate::visit

pub fn visit4<C1, C2, C3>( &self, visit: impl FnMut(InstanceKey, Primary, Option<C1>, Option<C2>, Option<C3>) ) -> Result<()>where C1: Component + Component, C2: Component + Component, C3: Component + Component, &'a C1: Into<Cow<'a, C1>> + 'a, &'a C2: Into<Cow<'a, C2>> + 'a, &'a C3: Into<Cow<'a, C3>> + 'a,

Visit the primary component of an EntityView. See crate::visit

pub fn visit5<C1, C2, C3, C4>( &self, visit: impl FnMut(InstanceKey, Primary, Option<C1>, Option<C2>, Option<C3>, Option<C4>) ) -> Result<()>where C1: Component + Component, C2: Component + Component, C3: Component + Component, C4: Component + Component, &'a C1: Into<Cow<'a, C1>> + 'a, &'a C2: Into<Cow<'a, C2>> + 'a, &'a C3: Into<Cow<'a, C3>> + 'a, &'a C4: Into<Cow<'a, C4>> + 'a,

Visit the primary component of an EntityView. See crate::visit

pub fn visit6<C1, C2, C3, C4, C5>( &self, visit: impl FnMut(InstanceKey, Primary, Option<C1>, Option<C2>, Option<C3>, Option<C4>, Option<C5>) ) -> Result<()>where C1: Component + Component, C2: Component + Component, C3: Component + Component, C4: Component + Component, C5: Component + Component, &'a C1: Into<Cow<'a, C1>> + 'a, &'a C2: Into<Cow<'a, C2>> + 'a, &'a C3: Into<Cow<'a, C3>> + 'a, &'a C4: Into<Cow<'a, C4>> + 'a, &'a C5: Into<Cow<'a, C5>> + 'a,

Visit the primary component of an EntityView. See crate::visit

pub fn visit7<C1, C2, C3, C4, C5, C6>( &self, visit: impl FnMut(InstanceKey, Primary, Option<C1>, Option<C2>, Option<C3>, Option<C4>, Option<C5>, Option<C6>) ) -> Result<()>where C1: Component + Component, C2: Component + Component, C3: Component + Component, C4: Component + Component, C5: Component + Component, C6: Component + Component, &'a C1: Into<Cow<'a, C1>> + 'a, &'a C2: Into<Cow<'a, C2>> + 'a, &'a C3: Into<Cow<'a, C3>> + 'a, &'a C4: Into<Cow<'a, C4>> + 'a, &'a C5: Into<Cow<'a, C5>> + 'a, &'a C6: Into<Cow<'a, C6>> + 'a,

Visit the primary component of an EntityView. See crate::visit

impl<'a, Primary> EntityView<Primary>where Primary: Component + 'a, &'a Primary: Into<Cow<'a, Primary>>,

pub fn as_df1(&self) -> Result<DataFrame>

pub fn as_df2<C1>(&self) -> Result<DataFrame>where C1: Component + 'a + Into<Cow<'a, C1>>,