Module re_query::range::results

Structs

RangeComponentResults

Lazily cached results for a particular component when using a cached range query.

RangeComponentResultsInner

Lazily cached results for a particular component when using a cached range query.

RangeData

RangeResults

Results for a range query.

Enums

Data 🔒

Helper datastructure to make it possible to convert latest-at results into ranged results.

Indices 🔒

Helper datastructure to make it possible to convert latest-at results into ranged results.

Constants

REENTERING 🔒

Keeps track of reentrancy counts for the current thread.