Function re_query::range_zip::range_zip_2x4

pub fn range_zip_2x4<Idx, IR0, R0, IR1, R1, IO0, O0, IO1, O1, IO2, O2, IO3, O3>(
    r0: IR0,
    r1: IR1,
    o0: IO0,
    o1: IO1,
    o2: IO2,
    o3: IO3
) -> RangeZip2x4<Idx, IR0::IntoIter, R0, IR1::IntoIter, R1, IO0::IntoIter, O0, IO1::IntoIter, O1, IO2::IntoIter, O2, IO3::IntoIter, O3> 
where
    Idx: Ord,
    IR0: IntoIterator<Item = (Idx, R0)>,
    IR1: IntoIterator<Item = (Idx, R1)>,
    IO0: IntoIterator<Item = (Idx, O0)>,
    IO1: IntoIterator<Item = (Idx, O1)>,
    IO2: IntoIterator<Item = (Idx, O2)>,
    IO3: IntoIterator<Item = (Idx, O3)>,

Returns a new RangeZip2x4 iterator.

The number of elements in a range zip iterator corresponds to the number of elements in the shortest of its required iterators (r0, r1).

Each call to next is guaranteed to yield the next value for each required iterator, as well as the most recent index amongst all of them.

Optional iterators accumulate their state and yield their most recent value (if any), each time the required iterators fire.