re_query::range_zip::generated
Function range_zip_1x1

pub fn range_zip_1x1<Idx, IR0, R0, IO0, O0>(
    r0: IR0,
    o0: IO0,
) -> RangeZip1x1<Idx, IR0::IntoIter, R0, IO0::IntoIter, O0> 
where
    Idx: Ord,
    IR0: IntoIterator<Item = (Idx, R0)>,
    IO0: IntoIterator<Item = (Idx, O0)>,

Returns a new RangeZip1x1 iterator.

The number of elements in a range zip iterator corresponds to the number of elements in the shortest of its required iterators (r0).

Each call to next is guaranteed to yield the next value for each required iterator, as well as the most recent index amongst all of them.

Optional iterators accumulate their state and yield their most recent value (if any), each time the required iterators fire.