Crate re_query

Provide query-centric access to the [re_data_store].

Modules

archetype_view 🔒

query 🔒

range 🔒

util 🔒

Structs

ArchetypeView

A view of an Archetype at a particular point in time returned by crate::get_component_with_instances.

ComponentNotFoundError

ComponentWithInstances

A type-erased array of Component values and the corresponding InstanceKey keys.

ExtraQueryHistory

When showing an entity in the history view, add this much history to it.

VisibleHistory

Visible history bounds.

Enums

QueryError

VisibleHistoryBoundary

One of the boundaries of the visible history.

Functions

get_component_with_instances

Retrieves a ComponentWithInstances from the [DataStore].

query_archetype

Retrieve an ArchetypeView from the DataStore, as well as the associated data time of its most recent component.

query_archetype_with_history

range_archetype

Iterates over the rows of any number of components and their respective cluster keys, all from the point-of-view of the required components, returning an iterator of ArchetypeViews.

range_component_set

Iterates over the rows of any number of components and their respective cluster keys, all from the point-of-view of the povs components, returning an iterator of ArchetypeViews.

Type Aliases

Result